+++ bc_obps/registration/tests/endpoints/user_operator/test_update_status.py
@@ -1,4 +1,3 @@
-from service.application_access_service import ApplicationAccessService
 from registration.enums.enums import AccessRequestStates, AccessRequestTypes
 from model_bakery import baker
 from registration.models import (
@@ -14,6 +13,7 @@ from registration.tests.utils.bakers import (
 )
 from registration.tests.utils.helpers import CommonTestSetup, TestUtils
 from registration.utils import custom_reverse_lazy
+from unittest.mock import patch
 
 
 class TestUserOperatorUpdateStatusEndpoint(CommonTestSetup):
@@ -43,13 +43,12 @@ class TestUserOperatorUpdateStatusEndpoint(CommonTestSetup):
         response_1_json = response_1.json()
         assert response_1_json == {"message": "Operator must be approved before approving or declining users."}
 
-    def test_industry_user_approves_access_request(self, mocker):
+    @patch('registration.middleware.registration_emails.send_operator_access_request_email')
+    def test_industry_user_approves_access_request(self, mock_send_operator_access_request_email):
         operator = operator_baker({"status": Operator.Statuses.APPROVED, "is_new": False})
         TestUtils.authorize_current_user_as_operator_user(self, operator=operator)
         subsequent_user_operator = baker.make(UserOperator, operator=operator)
-        mock_send_operator_access_request_email = mocker.patch.object(
-            ApplicationAccessService, "send_operator_access_request_email"
-        )
+
         response = TestUtils.mock_put_with_auth_role(
             self,
             "industry_user",
@@ -91,12 +90,11 @@ class TestUserOperatorUpdateStatusEndpoint(CommonTestSetup):
         )
         assert response.status_code == 403
 
-    def test_cas_admin_approves_access_request_with_existing_operator(self, mocker):
+    @patch('registration.middleware.registration_emails.send_operator_access_request_email')
+    def test_cas_admin_approves_access_request_with_existing_operator(self, mock_send_operator_access_request_email):
         operator = operator_baker({"status": Operator.Statuses.APPROVED, "is_new": False})
         user_operator = user_operator_baker({"operator": operator})
-        mock_send_operator_access_request_email = mocker.patch.object(
-            ApplicationAccessService, "send_operator_access_request_email"
-        )
+
         response_2 = TestUtils.mock_put_with_auth_role(
             self,
             "cas_admin",
@@ -122,14 +120,13 @@ class TestUserOperatorUpdateStatusEndpoint(CommonTestSetup):
             user_operator.user.email,
         )
 
-    def test_cas_admin_approves_admin_access_request_with_new_operator(self, mocker):
+    @patch('registration.middleware.registration_emails.send_operator_access_request_email')
+    def test_cas_admin_approves_admin_access_request_with_new_operator(self, mock_send_operator_access_request_email):
         # In this test we are testing the user operator status change and not the operator change,
         # so we have to mark the operator as is_new=False and status=APPROVED so we can bypass the below part and can get to the email sending part
         operator = operator_baker({'status': Operator.Statuses.APPROVED, 'is_new': False, 'created_by': self.user})
         user_operator = user_operator_baker({'operator': operator, 'user': operator.created_by})
-        mock_send_operator_access_request_email = mocker.patch.object(
-            ApplicationAccessService, "request_admin_access.send_operator_access_request_email", return_value=None
-        )
+
         response_2 = TestUtils.mock_put_with_auth_role(
             self,
             'cas_admin',
@@ -155,7 +152,8 @@ class TestUserOperatorUpdateStatusEndpoint(CommonTestSetup):
             user_operator.user.email,
         )
 
-    def test_cas_admin_declines_access_request(self, mocker):
+    @patch('registration.middleware.registration_emails.send_operator_access_request_email')
+    def test_cas_admin_declines_access_request(self, mock_send_operator_access_request_email):
         user = baker.make(User)
         operator = operator_baker()
         operator.status = Operator.Statuses.APPROVED
@@ -173,9 +171,7 @@ class TestUserOperatorUpdateStatusEndpoint(CommonTestSetup):
             business_role=BusinessRole.objects.get(role_name="Senior Officer"),
         )
         user_operator.operator.contacts.set(contacts)
-        mock_send_operator_access_request_email = mocker.patch.object(
-            ApplicationAccessService, "send_operator_access_request_email"
-        )
+
         # Now decline the user_operator and make sure the contacts are deleted
         response = TestUtils.mock_put_with_auth_role(
             self,
